Clarify “minimum system version requirements” summary
Change the phrase “you can only offer that update only to qualified users” to “you can restrict that update to qualified users.”

The former language implies that Sparkle prohibits offering the update to non-qualifying users, which is not accurate (until so configured). The intent is to convey that Sparkle can facilitate offering the update only to qualified users.